refactor(router): clarify routing host boundaries
