+++ lib/llm/src/kv_router.rs
@@ -50,8 +50,8 @@ pub mod indexer;
 pub mod metrics;
 pub mod prefill_router;
 pub mod publisher;
-pub mod push_router;
 mod route_lookup;
+mod routing_host;
 pub mod scheduler;
 pub mod sequence;
 pub mod shared_cache;
@@ -62,7 +62,8 @@ pub use dynamo_kv_router::scheduling::{
 pub use encoder_router::EncoderRouter;
 pub use indexer::{Indexer, ServedIndexerHandle, ServedIndexerMode, ensure_served_indexer_service};
 pub use prefill_router::PrefillRouter;
-pub use push_router::{BuiltinRoutingPolicy, KvPushRouter, RoutingHost};
+pub use routing_host::RoutingHost;
+pub(crate) use routing_host::{RoutingLoadState, builtin_policy_requires_load};
 
 use crate::{
     discovery::{KvSourceMembershipWatch, RuntimeConfigWatch},
